Initial check in for the Avaya mapping agent and discovery agent
Avaya Mapping agent gets called from Neutron ML2 plugin driver
whenever there is a port binding or port deletion operation to
carry out the actual provision of VLAN and ISIDs on the switch.
This mapping agent communicates with Avaya SDN Controller via
REST API to provision Avaya switches with appropriate ISID
and/or VLANs depending on the type of network being created.

Avaya Discovery agent uses LLDP to do dynamic discovery of the
network topology. This agent gathers information about the
switch ports to which the compute node and the network nodes
are connected. This information is then sent to the central
mapping agent. Mapping agent then utilizes this switch port
information to call SDN controller with the appropriate
parameters for the switch configuration.

Change-Id: I01f38d96a2c9a67a8afe43ee52d98d85f75d1a12